Separated classes and major definitions into new files
