+++ src/ui/party-ui-handler.ts
@@ -5,20 +5,17 @@ import { addBBCodeTextObject, addTextObject, getTextColor, TextStyle } from "#ap
 import { Command } from "#app/ui/command-ui-handler";
 import MessageUiHandler from "#app/ui/message-ui-handler";
 import { Mode } from "#app/ui/ui";
-import { BooleanHolder, toReadableString, randItem, getLocalizedSpriteKey } from "#app/utils";
+import { BooleanHolder, toReadableString, randItem } from "#app/utils";
 import {
   PokemonFormChangeItemModifier,
   PokemonHeldItemModifier,
   SwitchEffectTransferModifier,
 } from "#app/modifier/modifier";
 import { allMoves, ForceSwitchOutAttr } from "#app/data/move";
-import { getGenderColor, getGenderShadowColor, getGenderSymbol } from "#app/data/gender";
-import { StatusEffect } from "#enums/status-effect";
-import PokemonIconAnimHandler, { PokemonIconAnimMode } from "#app/ui/pokemon-icon-anim-handler";
+import PokemonIconAnimHandler from "#app/ui/pokemon-icon-anim-handler";
 import { pokemonEvolutions } from "#app/data/balance/pokemon-evolutions";
 import { addWindow } from "#app/ui/ui-theme";
 import { SpeciesFormChangeItemTrigger, FormChangeItem } from "#app/data/pokemon-forms";
-import { getVariantTint } from "#app/data/variant";
 import { Button } from "#enums/buttons";
 import { applyChallenges, ChallengeType } from "#app/data/challenge";
 import MoveInfoOverlay from "#app/ui/move-info-overlay";
@@ -32,119 +29,19 @@ import { SelectModifierPhase } from "#app/phases/select-modifier-phase";
 import { globalScene } from "#app/global-scene";
 import { determineMoveSource } from "./partyUtils/relearnMoveMode";
 import { MoveSource } from "#enums/move-source";
+import { PartySlot } from "./partyUtils/classes/party-slot";
+import { PartyCancelButton } from "./partyUtils/classes/party-cancel-button";
+import { PartyOption, PartyUiMode } from "#app/ui/partyUtils/definitions";
+import type {
+  PartySelectCallback,
+  PartyModifierSpliceSelectCallback,
+  PartyModifierTransferSelectCallback,
+  PokemonModifierTransferSelectFilter,
+  PokemonMoveSelectFilter,
+  PokemonSelectFilter,
+} from "#app/ui/partyUtils/definitions";
 
 const defaultMessage = i18next.t("partyUiHandler:choosePokemon");
-
-/**
- * Indicates the reason why the party UI is being opened.
- */
-export enum PartyUiMode {
-  /**
-   * Indicates that the party UI is open because of a user-opted switch.  This
-   * type of switch can be cancelled.
-   */
-  SWITCH,
-  /**
-   * Indicates that the party UI is open because of a faint or other forced
-   * switch (eg, move effect). This type of switch cannot be cancelled.
-   */
-  FAINT_SWITCH,
-  /**
-   * Indicates that the party UI is open because of a start-of-encounter optional
-   * switch. This type of switch can be cancelled.
-   */
-  POST_BATTLE_SWITCH,
-  /**
-   * Indicates that the party UI is open because of the move Revival Blessing.
-   * This selection cannot be cancelled.
-   */
-  REVIVAL_BLESSING,
-  /**
-   * Indicates that the party UI is open to select a mon to apply a modifier to.
-   * This type of selection can be cancelled.
-   */
-  MODIFIER,
-  /**
-   * Indicates that the party UI is open to select a mon to apply a move
-   * modifier to (such as an Ether or PP Up).  This type of selection can be cancelled.
-   */
-  MOVE_MODIFIER,
-  /**
-   * Indicates that the party UI is open to select a mon to teach a TM.  This
-   * type of selection can be cancelled.
-   */
-  TM_MODIFIER,
-  /**
-   * Indicates that the party UI is open to select a mon to remember a move.
-   * This type of selection can be cancelled.
-   */
-  REMEMBER_MOVE_MODIFIER,
-  /**
-   * Indicates that the party UI is open to transfer items between mons.  This
-   * type of selection can be cancelled.
-   */
-  MODIFIER_TRANSFER,
-  /**
-   * Indicates that the party UI is open because of a DNA Splicer.  This
-   * type of selection can be cancelled.
-   */
-  SPLICE,
-  /**
-   * Indicates that the party UI is open to release a party member.  This
-   * type of selection can be cancelled.
-   */
-  RELEASE,
-  /**
-   * Indicates that the party UI is open to check the team.  This
-   * type of selection can be cancelled.
-   */
-  CHECK,
-  /**
-   * Indicates that the party UI is open to select a party member for an arbitrary effect.
-   * This is generally used in for Mystery Encounter or special effects that require the player to select a Pokemon
-   */
-  SELECT,
-}
-
-export enum PartyOption {
-  CANCEL = -1,
-  SEND_OUT,
-  PASS_BATON,
-  REVIVE,
-  APPLY,
-  TEACH,
-  TRANSFER,
-  SUMMARY,
-  UNPAUSE_EVOLUTION,
-  SPLICE,
-  UNSPLICE,
-  RELEASE,
-  RENAME,
-  SELECT,
-  SCROLL_UP = 1000,
-  SCROLL_DOWN = 1001,
-  FORM_CHANGE_ITEM = 2000,
-  MOVE_1 = 3000,
-  MOVE_2,
-  MOVE_3,
-  MOVE_4,
-  ALL = 4000,
-}
-
-export type PartySelectCallback = (cursor: number, option: PartyOption) => void;
-export type PartyModifierTransferSelectCallback = (
-  fromCursor: number,
-  index: number,
-  itemQuantity?: number,
-  toCursor?: number,
-) => void;
-export type PartyModifierSpliceSelectCallback = (fromCursor: number, toCursor?: number) => void;
-export type PokemonSelectFilter = (pokemon: PlayerPokemon) => string | null;
-export type PokemonModifierTransferSelectFilter = (
-  pokemon: PlayerPokemon,
-  modifier: PokemonHeldItemModifier,
-) => string | null;
-export type PokemonMoveSelectFilter = (pokemonMove: PokemonMove) => string | null;
 
 export default class PartyUiHandler extends MessageUiHandler {
   private partyUiMode: PartyUiMode;
@@ -1024,33 +921,19 @@ export default class PartyUiHandler extends MessageUiHandler {
 
       for (let o = optionStartIndex; o < optionEndIndex; o++) {
         const option = this.options[this.options.length - (o + 1)];
-        let altTextColor = false;
+        // let altTextColor = false;
         let optionName: string = "";
         if ([PartyOption.CANCEL, PartyOption.SCROLL_UP, PartyOption.SCROLL_DOWN].includes(option)) {
-          switch (option) {
-            case PartyOption.CANCEL:
-              optionName = i18next.t(`partyUiHandler:${PartyOption[option]}`);
-              break;
-            case PartyOption.SCROLL_UP:
-              optionName = "↑";
-              break;
-            case PartyOption.SCROLL_DOWN:
-              optionName = "↓";
-              break;
-          }
+          optionName = this.getPartyOptionName(option);
         } else {
           optionName = allMoves[option].name;
           const moveSource = determineMoveSource(pokemon, allMoves[option].id);
           if (moveSource !== MoveSource.LEVEL_UP) {
-            altTextColor = true;
+            // altTextColor = true;
           }
         }
         const yCoord = -6 - 16 * o;
         const optionText = addBBCodeTextObject(0, yCoord - 16, optionName, TextStyle.WINDOW, { maxLines: 1 });
-        if (altTextColor) {
-          optionText.setColor("#40c8f8");
-          optionText.setShadowColor("#006090");
-        }
         optionText.setOrigin(0, 0);
         optionText.setText(`[shadow]${optionText.text}[/shadow]`);
 
@@ -1308,6 +1191,17 @@ export default class PartyUiHandler extends MessageUiHandler {
     }
   }
 
+  getPartyOptionName(option: PartyOption): string {
+    switch (option) {
+      case PartyOption.SCROLL_UP:
+        return "↑";
+      case PartyOption.SCROLL_DOWN:
+        return "↓";
+      default:
+        return i18next.t(`partyUiHandler:${PartyOption[option]}`);
+    }
+  }
+
   startTransfer(): void {
     this.transferMode = true;
     this.transferCursor = this.cursor;
@@ -1441,327 +1335,5 @@ export default class PartyUiHandler extends MessageUiHandler {
   clearPartySlots() {
     this.partySlots.splice(0, this.partySlots.length);
     this.partySlotsContainer.removeAll(true);
-  }
-}
-
-class PartySlot extends Phaser.GameObjects.Container {
-  private selected: boolean;
-  private transfer: boolean;
-  private slotIndex: number;
-  private pokemon: PlayerPokemon;
-
-  private slotBg: Phaser.GameObjects.Image;
-  private slotPb: Phaser.GameObjects.Sprite;
-  public slotName: Phaser.GameObjects.Text;
-  public slotHpBar: Phaser.GameObjects.Image;
-  public slotHpOverlay: Phaser.GameObjects.Sprite;
-  public slotHpText: Phaser.GameObjects.Text;
-  public slotDescriptionLabel: Phaser.GameObjects.Text; // this is used to show text instead of the HP bar i.e. for showing "Able"/"Not Able" for TMs when you try to learn them
-
-  private pokemonIcon: Phaser.GameObjects.Container;
-  private iconAnimHandler: PokemonIconAnimHandler;
-
-  constructor(
-    slotIndex: number,
-    pokemon: PlayerPokemon,
-    iconAnimHandler: PokemonIconAnimHandler,
-    partyUiMode: PartyUiMode,
-    tmMoveId: Moves,
-  ) {
-    super(
-      globalScene,
-      slotIndex >= globalScene.currentBattle.getBattlerCount() ? 230.5 : 64,
-      slotIndex >= globalScene.currentBattle.getBattlerCount()
-        ? -184
-            + (globalScene.currentBattle.double ? -40 : 0)
-            + (28 + (globalScene.currentBattle.double ? 8 : 0)) * slotIndex
-        : -124 + (globalScene.currentBattle.double ? -8 : 0) + slotIndex * 64,
-    );
-
-    this.slotIndex = slotIndex;
-    this.pokemon = pokemon;
-    this.iconAnimHandler = iconAnimHandler;
-
-    this.setup(partyUiMode, tmMoveId);
-  }
-
-  getPokemon(): PlayerPokemon {
-    return this.pokemon;
-  }
-
-  setup(partyUiMode: PartyUiMode, tmMoveId: Moves) {
-    const battlerCount = globalScene.currentBattle.getBattlerCount();
-
-    const slotKey = `party_slot${this.slotIndex >= battlerCount ? "" : "_main"}`;
-
-    const slotBg = globalScene.add.sprite(0, 0, slotKey, `${slotKey}${this.pokemon.hp ? "" : "_fnt"}`);
-    this.slotBg = slotBg;
-
-    this.add(slotBg);
-
-    const slotPb = globalScene.add.sprite(
-      this.slotIndex >= battlerCount ? -85.5 : -51,
-      this.slotIndex >= battlerCount ? 0 : -20.5,
-      "party_pb",
-    );
-    this.slotPb = slotPb;
-
-    this.add(slotPb);
-
-    this.pokemonIcon = globalScene.addPokemonIcon(this.pokemon, slotPb.x, slotPb.y, 0.5, 0.5, true);
-
-    this.add(this.pokemonIcon);
-
-    this.iconAnimHandler.addOrUpdate(this.pokemonIcon, PokemonIconAnimMode.PASSIVE);
-
-    const slotInfoContainer = globalScene.add.container(0, 0);
-    this.add(slotInfoContainer);
-
-    let displayName = this.pokemon.getNameToRender();
-    let nameTextWidth: number;
-
-    const nameSizeTest = addTextObject(0, 0, displayName, TextStyle.PARTY);
-    nameTextWidth = nameSizeTest.displayWidth;
-
-    while (nameTextWidth > (this.slotIndex >= battlerCount ? 52 : 76 - (this.pokemon.fusionSpecies ? 8 : 0))) {
-      displayName = `${displayName.slice(0, displayName.endsWith(".") ? -2 : -1).trimEnd()}.`;
-      nameSizeTest.setText(displayName);
-      nameTextWidth = nameSizeTest.displayWidth;
-    }
-
-    nameSizeTest.destroy();
-
-    this.slotName = addTextObject(0, 0, displayName, TextStyle.PARTY);
-    this.slotName.setPositionRelative(
-      slotBg,
-      this.slotIndex >= battlerCount ? 21 : 24,
-      this.slotIndex >= battlerCount ? 2 : 10,
-    );
-    this.slotName.setOrigin(0, 0);
-
-    const slotLevelLabel = globalScene.add.image(0, 0, "party_slot_overlay_lv");
-    slotLevelLabel.setPositionRelative(this.slotName, 8, 12);
-    slotLevelLabel.setOrigin(0, 0);
-
-    const slotLevelText = addTextObject(
-      0,
-      0,
-      this.pokemon.level.toString(),
-      this.pokemon.level < globalScene.getMaxExpLevel() ? TextStyle.PARTY : TextStyle.PARTY_RED,
-    );
-    slotLevelText.setPositionRelative(slotLevelLabel, 9, 0);
-    slotLevelText.setOrigin(0, 0.25);
-
-    slotInfoContainer.add([this.slotName, slotLevelLabel, slotLevelText]);
-
-    const genderSymbol = getGenderSymbol(this.pokemon.getGender(true));
-
-    if (genderSymbol) {
-      const slotGenderText = addTextObject(0, 0, genderSymbol, TextStyle.PARTY);
-      slotGenderText.setColor(getGenderColor(this.pokemon.getGender(true)));
-      slotGenderText.setShadowColor(getGenderShadowColor(this.pokemon.getGender(true)));
-      if (this.slotIndex >= battlerCount) {
-        slotGenderText.setPositionRelative(slotLevelLabel, 36, 0);
-      } else {
-        slotGenderText.setPositionRelative(this.slotName, 76 - (this.pokemon.fusionSpecies ? 8 : 0), 3);
-      }
-      slotGenderText.setOrigin(0, 0.25);
-
-      slotInfoContainer.add(slotGenderText);
-    }
-
-    if (this.pokemon.fusionSpecies) {
-      const splicedIcon = globalScene.add.image(0, 0, "icon_spliced");
-      splicedIcon.setScale(0.5);
-      splicedIcon.setOrigin(0, 0);
-      if (this.slotIndex >= battlerCount) {
-        splicedIcon.setPositionRelative(slotLevelLabel, 36 + (genderSymbol ? 8 : 0), 0.5);
-      } else {
-        splicedIcon.setPositionRelative(this.slotName, 76, 3.5);
-      }
-
-      slotInfoContainer.add(splicedIcon);
-    }
-
-    if (this.pokemon.status) {
-      const statusIndicator = globalScene.add.sprite(0, 0, getLocalizedSpriteKey("statuses"));
-      statusIndicator.setFrame(StatusEffect[this.pokemon.status?.effect].toLowerCase());
-      statusIndicator.setOrigin(0, 0);
-      statusIndicator.setPositionRelative(slotLevelLabel, this.slotIndex >= battlerCount ? 43 : 55, 0);
-
-      slotInfoContainer.add(statusIndicator);
-    }
-
-    if (this.pokemon.isShiny()) {
-      const doubleShiny = this.pokemon.isFusion() && this.pokemon.shiny && this.pokemon.fusionShiny;
-
-      const shinyStar = globalScene.add.image(0, 0, `shiny_star_small${doubleShiny ? "_1" : ""}`);
-      shinyStar.setOrigin(0, 0);
-      shinyStar.setPositionRelative(this.slotName, -9, 3);
-      shinyStar.setTint(getVariantTint(!doubleShiny ? this.pokemon.getVariant() : this.pokemon.variant));
-
-      slotInfoContainer.add(shinyStar);
-
-      if (doubleShiny) {
-        const fusionShinyStar = globalScene.add.image(0, 0, "shiny_star_small_2");
-        fusionShinyStar.setOrigin(0, 0);
-        fusionShinyStar.setPosition(shinyStar.x, shinyStar.y);
-        fusionShinyStar.setTint(getVariantTint(this.pokemon.fusionVariant));
-
-        slotInfoContainer.add(fusionShinyStar);
-      }
-    }
-
-    this.slotHpBar = globalScene.add.image(0, 0, "party_slot_hp_bar");
-    this.slotHpBar.setPositionRelative(
-      slotBg,
-      this.slotIndex >= battlerCount ? 72 : 8,
-      this.slotIndex >= battlerCount ? 6 : 31,
-    );
-    this.slotHpBar.setOrigin(0, 0);
-    this.slotHpBar.setVisible(false);
-
-    const hpRatio = this.pokemon.getHpRatio();
-
-    this.slotHpOverlay = globalScene.add.sprite(
-      0,
-      0,
-      "party_slot_hp_overlay",
-      hpRatio > 0.5 ? "high" : hpRatio > 0.25 ? "medium" : "low",
-    );
-    this.slotHpOverlay.setPositionRelative(this.slotHpBar, 16, 2);
-    this.slotHpOverlay.setOrigin(0, 0);
-    this.slotHpOverlay.setScale(hpRatio, 1);
-    this.slotHpOverlay.setVisible(false);
-
-    this.slotHpText = addTextObject(0, 0, `${this.pokemon.hp}/${this.pokemon.getMaxHp()}`, TextStyle.PARTY);
-    this.slotHpText.setPositionRelative(this.slotHpBar, this.slotHpBar.width - 3, this.slotHpBar.height - 2);
-    this.slotHpText.setOrigin(1, 0);
-    this.slotHpText.setVisible(false);
-
-    this.slotDescriptionLabel = addTextObject(0, 0, "", TextStyle.MESSAGE);
-    this.slotDescriptionLabel.setPositionRelative(
-      slotBg,
-      this.slotIndex >= battlerCount ? 94 : 32,
-      this.slotIndex >= battlerCount ? 16 : 46,
-    );
-    this.slotDescriptionLabel.setOrigin(0, 1);
-    this.slotDescriptionLabel.setVisible(false);
-
-    slotInfoContainer.add([this.slotHpBar, this.slotHpOverlay, this.slotHpText, this.slotDescriptionLabel]);
-
-    if (partyUiMode !== PartyUiMode.TM_MODIFIER) {
-      this.slotDescriptionLabel.setVisible(false);
-      this.slotHpBar.setVisible(true);
-      this.slotHpOverlay.setVisible(true);
-      this.slotHpText.setVisible(true);
-    } else {
-      this.slotHpBar.setVisible(false);
-      this.slotHpOverlay.setVisible(false);
-      this.slotHpText.setVisible(false);
-      let slotTmText: string;
-
-      if (this.pokemon.getMoveset().filter((m) => m.moveId === tmMoveId).length > 0) {
-        slotTmText = i18next.t("partyUiHandler:learned");
-      } else if (this.pokemon.compatibleTms.indexOf(tmMoveId) === -1) {
-        slotTmText = i18next.t("partyUiHandler:notAble");
-      } else {
-        slotTmText = i18next.t("partyUiHandler:able");
-      }
-
-      this.slotDescriptionLabel.setText(slotTmText);
-      this.slotDescriptionLabel.setVisible(true);
-    }
-  }
-
-  select(): void {
-    if (this.selected) {
-      return;
-    }
-
-    this.selected = true;
-    this.iconAnimHandler.addOrUpdate(this.pokemonIcon, PokemonIconAnimMode.ACTIVE);
-
-    this.updateSlotTexture();
-    this.slotPb.setFrame("party_pb_sel");
-  }
-
-  deselect(): void {
-    if (!this.selected) {
-      return;
-    }
-
-    this.selected = false;
-    this.iconAnimHandler.addOrUpdate(this.pokemonIcon, PokemonIconAnimMode.PASSIVE);
-
-    this.updateSlotTexture();
-    this.slotPb.setFrame("party_pb");
-  }
-
-  setTransfer(transfer: boolean): void {
-    if (this.transfer === transfer) {
-      return;
-    }
-
-    this.transfer = transfer;
-    this.updateSlotTexture();
-  }
-
-  private updateSlotTexture(): void {
-    const battlerCount = globalScene.currentBattle.getBattlerCount();
-    this.slotBg.setTexture(
-      `party_slot${this.slotIndex >= battlerCount ? "" : "_main"}`,
-      `party_slot${this.slotIndex >= battlerCount ? "" : "_main"}${this.transfer ? "_swap" : this.pokemon.hp ? "" : "_fnt"}${this.selected ? "_sel" : ""}`,
-    );
-  }
-}
-
-class PartyCancelButton extends Phaser.GameObjects.Container {
-  private selected: boolean;
-
-  private partyCancelBg: Phaser.GameObjects.Sprite;
-  private partyCancelPb: Phaser.GameObjects.Sprite;
-
-  constructor(x: number, y: number) {
-    super(globalScene, x, y);
-
-    this.setup();
-  }
-
-  setup() {
-    const partyCancelBg = globalScene.add.sprite(0, 0, "party_cancel");
-    this.add(partyCancelBg);
-
-    this.partyCancelBg = partyCancelBg;
-
-    const partyCancelPb = globalScene.add.sprite(-17, 0, "party_pb");
-    this.add(partyCancelPb);
-
-    this.partyCancelPb = partyCancelPb;
-
-    const partyCancelText = addTextObject(-8, -7, i18next.t("partyUiHandler:cancel"), TextStyle.PARTY);
-    this.add(partyCancelText);
-  }
-
-  select() {
-    if (this.selected) {
-      return;
-    }
-
-    this.selected = true;
-
-    this.partyCancelBg.setFrame("party_cancel_sel");
-    this.partyCancelPb.setFrame("party_pb_sel");
-  }
-
-  deselect() {
-    if (!this.selected) {
-      return;
-    }
-
-    this.selected = false;
-
-    this.partyCancelBg.setFrame("party_cancel");
-    this.partyCancelPb.setFrame("party_pb");
   }
 }
