+++ src/ui/partyUtils/classes/party-cancel-button.ts
@@ -1,0 +1,53 @@
+import { globalScene } from "#app/global-scene";
+import { addTextObject, TextStyle } from "../../text";
+import i18next from "i18next";
+
+export class PartyCancelButton extends Phaser.GameObjects.Container {
+  private selected: boolean;
+
+  private partyCancelBg: Phaser.GameObjects.Sprite;
+  private partyCancelPb: Phaser.GameObjects.Sprite;
+
+  constructor(x: number, y: number) {
+    super(globalScene, x, y);
+
+    this.setup();
+  }
+
+  setup() {
+    const partyCancelBg = globalScene.add.sprite(0, 0, "party_cancel");
+    this.add(partyCancelBg);
+
+    this.partyCancelBg = partyCancelBg;
+
+    const partyCancelPb = globalScene.add.sprite(-17, 0, "party_pb");
+    this.add(partyCancelPb);
+
+    this.partyCancelPb = partyCancelPb;
+
+    const partyCancelText = addTextObject(-8, -7, i18next.t("partyUiHandler:cancel"), TextStyle.PARTY);
+    this.add(partyCancelText);
+  }
+
+  select() {
+    if (this.selected) {
+      return;
+    }
+
+    this.selected = true;
+
+    this.partyCancelBg.setFrame("party_cancel_sel");
+    this.partyCancelPb.setFrame("party_pb_sel");
+  }
+
+  deselect() {
+    if (!this.selected) {
+      return;
+    }
+
+    this.selected = false;
+
+    this.partyCancelBg.setFrame("party_cancel");
+    this.partyCancelPb.setFrame("party_pb");
+  }
+}
